refactor(application): make the proper use of the application service/usecase
