--- src/core/domain/model/numbers.ts
@@ -1,42 +1,0 @@
-type Data = string | { [n: number]: string };
-
-export const units: Data = {
-  0: 'zero',
-  1: 'one',
-  2: 'two',
-  3: 'three',
-  4: 'four',
-  5: 'five',
-  6: 'six',
-  7: 'seven',
-  8: 'eight',
-  9: 'nine',
-};
-
-export const specialDozens: Data = {
-  10: 'ten',
-  11: 'eleven',
-  12: 'twelve',
-  13: 'thirteen',
-  14: 'fourteen',
-  15: 'fifteen',
-  16: 'sixteen',
-  17: 'seventeen',
-  18: 'eighteen',
-  19: 'nineteen',
-};
-
-export const dozens: Data = {
-  20: 'twenty',
-  30: 'thirty',
-  40: 'forty',
-  50: 'fifty',
-  60: 'sixty',
-  70: 'seventy',
-  80: 'eighty',
-  90: 'ninety',
-};
-
-export const hundred: Data = 'hundred';
-
-export const thousand: Data = 'thousand';
