--- src/core/domain/protocols/convert-number-to-full-text.ts
@@ -1,3 +1,0 @@
-export interface ConvertNumberToFullText {
-  invoke(number: number): string;
-}
