--- src/core/usecases/convert-number-to-full-text/index.ts
@@ -1,1 +1,0 @@
-export * from './convert-number-to-full-text-usecase';
